+++ common/src/main/java/dev/gimme/sharedlife/domain/config/ServerConfig.java
@@ -7,4 +7,6 @@ public abstract class ServerConfig {
     public abstract boolean shareHealth();
     public abstract boolean shareHunger();
     public abstract boolean shareExperience();
+    public abstract boolean announceDamage();
+    public abstract boolean includeDamageSource();
 }