+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add option to disable damage announcements in chat
+
 ## [v1.21.1-0.1.0] - 2026-01-07
 
 ### Added